Delauney attempt 1

+++ drone_demo-main/vector2d.cpp
@@ -1,10 +1,10 @@
 #include "vector2d.h"
 
 const Vector2D operator *(double a,const Vector2D &v) {
-    return Vector2D(float(a*v.x),float(a*v.y));
+    return Vector2D(a*v.x,a*v.y);
 }
 
- double operator *(const Vector2D &u,const Vector2D &v) {
+double operator *(const Vector2D &u,const Vector2D &v) {
     return u.x*v.x+u.y*v.y;
 }
 
@@ -20,10 +20,10 @@ const Vector2D operator-(const Vector2D &v) {
     return Vector2D(-v.x,-v.y);
 }
 
- double operator ^(const Vector2D &u,const Vector2D &v) {
+double operator ^(const Vector2D &u,const Vector2D &v) {
     return u.x*v.y-u.y*v.x;
 }
 
- bool operator==(const Vector2D &u,const Vector2D &v) {
+bool operator==(const Vector2D &u,const Vector2D &v) {
     return (u.x==v.x && u.y==v.y);
 }
